Improve communication between all processes.

+++ src/EloGank/Api/Client/Async/LOLAsyncClient.php
@@ -5,6 +5,7 @@ namespace EloGank\Api\Client\Async;
 use EloGank\Api\Client\LOLClientInterface;
 use EloGank\Api\Component\Configuration\ConfigurationLoader;
 use EloGank\Api\Logger\LoggerFactory;
+use EloGank\Api\Model\Region\RegionInterface;
 use Predis\Client;
 use Psr\Log\LoggerInterface;
 
@@ -24,6 +25,21 @@ class LOLAsyncClient implements LOLClientInterface
     protected $clientId;
 
     /**
+     * @var RegionInterface
+     */
+    protected $region;
+
+    /**
+     * @var int
+     */
+    protected $port;
+
+    /**
+     * @var string
+     */
+    protected $error;
+
+    /**
      * @var \ZMQSocket
      */
     protected $con;
@@ -40,26 +56,35 @@ class LOLAsyncClient implements LOLClientInterface
 
 
     /**
-     * @param int $accountKey
-     * @param int $clientId
+     * @param Client          $redis
+     * @param int             $accountKey
+     * @param int             $clientId
+     * @param RegionInterface $region
+     * @param int             $port
      */
-    public function __construct($accountKey, $clientId)
+    public function __construct(Client $redis, $accountKey, $clientId, RegionInterface $region, $port)
     {
         $rootFolder = __DIR__ . '/../../../../..';
 
         $this->pidPath  = $rootFolder . '/' . ConfigurationLoader::get('cache.path') . '/client_' . $clientId . '.pid';
+        $this->redis    = $redis;
         $this->clientId = $clientId;
+        $this->region   = $region;
+        $this->port     = $port;
+
         $this->con      = new \ZMQSocket(new \ZMQContext(), \ZMQ::SOCKET_PUSH);
-        $this->redis    = new Client('tcp://127.0.0.1:6379');
         $this->logger   = LoggerFactory::create();
 
         // Create process
         popen(sprintf('php %s/console elogank:client:create %d %d > /dev/null 2>&1 & echo $! > %s', $rootFolder, $accountKey, $clientId, $this->pidPath), 'r');
     }
 
+    /**
+     * Authenticate the client
+     */
     public function authenticate()
     {
-        $this->con->connect('tcp://127.0.0.1:5555');
+        $this->con->connect('tcp://127.0.0.1:' . $this->port);
 
         $this->send('authenticate');
     }
@@ -74,6 +99,8 @@ class LOLAsyncClient implements LOLClientInterface
             return json_decode($message[1], true)['is_authenticated'];
         }
 
+        $this->error = 'Asynchronous client did not respond';
+
         return false;
     }
 
@@ -85,14 +112,28 @@ class LOLAsyncClient implements LOLClientInterface
         return $this->clientId;
     }
 
+    /**
+     * @return RegionInterface
+     */
     public function getRegion()
     {
-        // TODO: Implement getRegion() method.
+        return $this->region->getUniqueName();
     }
 
+    /**
+     * @return string
+     */
     public function getError()
     {
-        // TODO: Implement getError() method.
+        return $this->error;
+    }
+
+    /**
+     * @return int
+     */
+    public function getPort()
+    {
+        return $this->port;
     }
 
     /**
@@ -100,6 +141,11 @@ class LOLAsyncClient implements LOLClientInterface
      */
     public function kill()
     {
+        // Client doesn't exist
+        if (!is_file($this->pidPath)) {
+            return;
+        }
+
         $pid = file_get_contents($this->pidPath);
         unlink($this->pidPath);
 
@@ -122,5 +168,4 @@ class LOLAsyncClient implements LOLClientInterface
             'parameters' => $parameters
         ]));
     }
-
 }
